requirements.txt
# Web framework
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.3

# PDF processing
PyMuPDF==1.23.22  # fitz

# Embedding and search
faiss-cpu==1.7.4

# Google Gemini API
google-generativeai==0.5.0

# CrewAI orchestration
crewai==0.28.8

# Data handling
numpy==1.26.4

# Optional: CORS for frontend/backend communication
fastapi[all]
python-dotenv
httpx 
beautifulsoup4
